+++ src/com/topodroid/Cave3D/Cave3DIntersection.java
@@ -15,15 +15,15 @@ package com.topodroid.Cave3D;
 // 3D intersection point on the segment
 class Cave3DIntersection extends Vector3D
 {
-  float s; // line abscissa on the shot P(s) = p1 + s ( p2-p1)
+  double s; // line abscissa on the shot P(s) = p1 + s ( p2-p1)
  
-  Cave3DIntersection( Vector3D pt, float s0 )
+  Cave3DIntersection( Vector3D pt, double s0 )
   {
     super( pt );
     s = s0;
   }
 
-  Cave3DIntersection( float x0, float y0, float z0, float s0 )
+  Cave3DIntersection( double x0, double y0, double z0, double s0 )
   {
     super( x0, y0, z0 );
     s = s0;
