+++ src/com/topodroid/Cave3D/Cave3DPolygon.java
@@ -40,11 +40,11 @@ class Cave3DPolygon
     return false;
   }
 
-  float getAverageDistance() 
+  double getAverageDistance() 
   {
     int ns = points.size();
-    if ( ns <= 1 ) return 0.0f;
-    float d = 0.0f;
+    if ( ns <= 1 ) return 0.0;
+    double d = 0.0;
     for ( int k=1; k<ns; ++k ) d += Vector3D.distance3D( points.get(k-1), points.get(k) );
     return d / (ns - 1);
   }
